refactor: partner, layout api 타입 분리
