+++ apps/client/lib/api/layout.types.ts
@@ -1,0 +1,8 @@
+export interface ExhibitionFooter {
+	title: string;
+	department: string;
+	address: string | null;
+	detail_location: string | null;
+	email: string;
+	copyright: string | null;
+}
