+++ apps/client/lib/api/partner.ts
@@ -1,21 +1,7 @@
-// api/partner.ts
-
 import { apiClient } from "api";
+import type { PartnerPart } from "./partner.types";
 
-export interface PartnerMember {
-	member_id: string;
-	member_name: string;
-	member_name_en?: string;
-	member_email?: string;
-	member_image_url?: string;
-}
-
-export interface PartnerPart {
-	order: number;
-	part_id: string;
-	part_name: string;
-	members: PartnerMember[];
-}
+export type * from "./partner.types";
 
 interface GetPartnersResponse {
 	parts: PartnerPart[];
@@ -35,7 +21,7 @@ export async function getPartners(exhibitionId: string): Promise<PartnerPart[]> 
 		}));
 
 		return sortedParts;
-	} catch (_error) {
+	} catch {
 		return [];
 	}
 }
